syssstat: Implement click_event handling
This allows us to receive events from i3 when the user clicks on the
status bar. For now, all it does is switch between fuzzy and "precise"
clock when clicking on any part of the status bar, but everything is in
place to extend this even further.

Implementing the JSON stream-parsing was quite the pain, I'm glad I
finally have it done now.